+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `NM_DATE_FORMAT_LONG` constant for long-form date display (#499)
+
+### Changed
+
+- Migrate `'j F Y'` literals across templates to `NM_DATE_FORMAT_LONG`
+
 ## [4.5.5] - 2026-04-27
 
 ### Added